chore(SystemsTable): Rename InventoryTable component
