+++ src/SmartComponents/index.js
@@ -4,4 +4,4 @@ export { default as CreatePolicy } from './CreatePolicy/CreatePolicy';
 export { default as DeleteReport } from './DeleteReport/DeleteReport';
 export { default as EditPolicy } from './EditPolicy/EditPolicy';
 export { default as DeletePolicy } from './DeletePolicy/DeletePolicy';
-export { default as InventoryTable } from './SystemsTable/InventoryTable';
+export { default as SystemsTable } from './SystemsTable/SystemsTable';
